Add files via upload
Changes have been made to these files to allow the user to switch easily between the "standard" 4-pin unipolar drives, as used by the Openflexure microscope, and 2-pin bipolar drives. The latter come in a wider range of sizes and power, allowing this code to have a wider application range. It has been tested using the DRV8825 driver module.